+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-28T23:42:18.750Z",
+    "2026-05-29T05:33:39.324Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {